Adding the new "edit_circle" icon to the Forge icon library
For the Report Writing project in Public Safety we have a need for a new icon to be added to meet the style that is being used on our complex stepper states. I dropped the "edit_circle.svg" into SVG>Custom per Emily Rollers guidance. The icon has been optimized and is void of any type of inline styling.